+++ practice/orderingSystem/c11g/Product.js
@@ -1,0 +1,13 @@
+class Product {
+  constructor(id,name,stock){
+    this.id = id;
+    this.name = name;
+    this.stock = stock;
+  }
+
+  setStock(count){
+    this.stock = count;
+  }
+}
+
+export default Product;
